requirements/requirements-document-chatbot.txt
chromadb~=0.4.22
langchain~=0.1.0
langchain_community==0.0.12
python-dotenv~=1.0.0
streamlit~=1.30.0
openai~=1.7.2
llama-cpp-python~=0.2.28
unstructured~=0.12.0
python-docx~=1.1.0
